Initially restructured file hierarchy; moved deprecations.
The file hierarchy has been restructured to contain all data stored in
Pmw/Pmw_X_Y_Z/ beforehand in the main directory. The leading Pmw/ directory
has been removed from the path as it is deemed to be unnecessary.

bundlepmw.py (in bin/), __init__.py (Pmw global init), maketar.sh, README,
set_env.sh are considered to be outdated, they are deprecated or shall be
replaced by a recent version. This files are kept in the deprecated/
directory.