Rename of Pmw library files.
The Pmw lib files have been renamed from the non-standard naming convention
AaaaBbbb.py to the standardised convention aaaa_bbbb.py. This will make

>>> import pmw2.about_dlg

and further

>>> pmw2.about_dlg.AboutDlg

possible.  It is a cleaner way of specifying what to import in derived
modules than to import Pmw at all and trusting that the class desired is
already imported.