First working version of pmw2.
This version works, but is not yet backward compatible with
Pmw regarding attribute access.

The dynload module has been fixed and is available now for
dynamic loading of modules.  Missing is the dynamic loading
of module attributes.  This is needed to achieve attributes
of pmw2 like pmw2.ScrolledFrame, which reference an object
in a module to be dynloaded.

about_dlg.py has been renamed to about_dialog.py, as the
_dialog suffix is used in prompt_dialog.py and company too.

loader.py has been moved to deprecated, as it belongs there.

The modules still with an uppercase letter in the beginning
in the Git repo have been renamed to start lowercase.

import statements have again been modified, as pmw2 supplies
its submodules via attributes which are instances of
pmw2.dynload.Dynload.  Thus directly importing e.g. pmw2.base
in some module is not necessary and contraproductive.