Beta stage reached.
This is a version of pmw2, which is intended to be fully backward
compatible.  Thereby dynload is maintained, and it should be compatible
with py2exe (and py2app).  First test indicate this.  Further tests are
in the queue.

dynload.py has been updated to support the dynload of module attributes.
__init__.py has been updated to dynload all the publics of the modules.
base.py has been updated to use pmw2 for the import of logical_fonts in
initialise().